bcqm_tools/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bcqm-tools-local"
version = "0.2.0"
description = "Reproducible tools for BCQM (Ramsey/W bound, GKLS qubit, Galton)."
readme = "README.md"
authors = [{name="Peter M. Ferguson"}, {name="ChatGPT (assistant)"}]
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.22",
    "scipy>=1.10",
    "matplotlib>=3.7",
    "pyyaml>=6.0",
]

[project.scripts]
bcqm = "bcqm_tools.cli:main"

[tool.setuptools]
packages = ["bcqm_tools"]
